gblog/requirements.txt
dj-database-url==0.3.0
Django==1.7.5
django-appconf==1.0.1
django-compressor==1.4
django-filter==0.9.2
django-taggit==0.12.3
djangorestframework==3.1.0
drf-nested-routers==0.9.0
pillow==2.7.0
six==1.9.0
